Make discounts truly per-product + add tax feature

+++ database/migrations/2019_07_06_153441_create_payments_table.php
@@ -17,7 +17,7 @@ class CreatePaymentsTable extends Migration
             $table->integer('responsable_id')->unsigned();
             $table->integer('pre_invoice_id')->unsigned();
             $table->string('payment_method')->nullable();
-            $table->decimal('value', 8, 2);
+            $table->bigInteger('value');
             $table->string('comment')->nullable();
             $table->timestamps();
             //$table->softDeletes();
